neural/pyproject.toml
[project]
name = "fianchetto-neural"
version = "1.0.0"
description = "Neural evaluation service for Fianchetto Engine"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-chess>=1.999",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

